// frontend/src/api/models/OverviewWahlkreis.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Donator-Mainframe communication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Abgeordneter } from './Abgeordneter';
import {
    AbgeordneterFromJSON,
    AbgeordneterFromJSONTyped,
    AbgeordneterToJSON,
    AbgeordneterToJSONTyped,
} from './Abgeordneter';

/**
 * 
 * @export
 * @interface OverviewWahlkreis
 */
export interface OverviewWahlkreis {
    /**
     * 
     * @type {number}
     * @memberof OverviewWahlkreis
     */
    wahlbeteiligung?: number;
    /**
     * 
     * @type {Abgeordneter}
     * @memberof OverviewWahlkreis
     */
    direktkandidat?: Abgeordneter;
}

/**
 * Check if a given object implements the OverviewWahlkreis interface.
 */
export function instanceOfOverviewWahlkreis(value: object): value is OverviewWahlkreis {
    return true;
}

export function OverviewWahlkreisFromJSON(json: any): OverviewWahlkreis {
    return OverviewWahlkreisFromJSONTyped(json, false);
}

export function OverviewWahlkreisFromJSONTyped(json: any, ignoreDiscriminator: boolean): OverviewWahlkreis {
    if (json == null) {
        return json;
    }
    return {
        
        'wahlbeteiligung': json['wahlbeteiligung'] == null ? undefined : json['wahlbeteiligung'],
        'direktkandidat': json['direktkandidat'] == null ? undefined : AbgeordneterFromJSON(json['direktkandidat']),
    };
}

export function OverviewWahlkreisToJSON(json: any): OverviewWahlkreis {
    return OverviewWahlkreisToJSONTyped(json, false);
}

export function OverviewWahlkreisToJSONTyped(value?: OverviewWahlkreis | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'wahlbeteiligung': value['wahlbeteiligung'],
        'direktkandidat': AbgeordneterToJSON(value['direktkandidat']),
    };
}